Crate relm4_macros

A collection of macros for gtk-rs, Relm4 and Rust in general.

The docs are available in two versions. Use the stable docs if you want get information about a version that was already published. Visit the nightly docs if are trying out the newest but possibly unstable version of the crate.

Docs of related crates: relm4 | relm4-macros | relm4-components | gtk4-rs | gtk-rs-core | libadwaita-rs | libpanel-rs

GitHub | Website | Book | Blog

Macros

menu

A macro to create menus.

view

The view! macro allows you to construct your UI easily and cleanly.

Attribute Macros

component

Macro that implements relm4::Component or relm4::SimpleComponent and generates the corresponding widgets struct.

factory

Macro that implements relm4::factory::FactoryComponent and generates the corresponding widgets struct.

widget_template

A macro to generate widget templates.